* reset innermarkingcount_t to uint32_t * the FullMarkingQueue now adaptively reserves memory based on the past

+++ wendy/src/types.h
@@ -33,7 +33,7 @@ typedef uint16_t hash_t;
 typedef uint8_t Label_ID;
 
 /// the range of markings (all markings, deadlock markings) stored within a knowledge (max. 65535)
-typedef uint16_t innermarkingcount_t;
+typedef uint32_t innermarkingcount_t;
 
 /// the range of inner markings (max. 4294967296)
 typedef uint32_t InnerMarking_ID;
